Rename old WorkflowStatusEditor to GranularWorkflowStatusEditor
