+++ src/RDF/Editor/GraphEditor.php
@@ -49,7 +49,7 @@ final class GraphEditor
 
         // Set the udb:workflowStatus property to draft if not set yet.
         // TODO: This can be removed once everything is ported to JSON conversion to turtle.
-        WorkflowStatusEditor::for($this->graph)->draft($resourceIri);
+        GranularWorkflowStatusEditor::for($this->graph)->draft($resourceIri);
 
         // Set the dcterms:created property if not set yet.
         // (Otherwise it would constantly update like dcterms:modified).
