+++ src/RDF/Editor/WorkflowStatusEditor.php
@@ -6,14 +6,11 @@ namespace CultuurNet\UDB3\RDF\Editor;
 
 use CultuurNet\UDB3\Model\ValueObject\Moderation\WorkflowStatus;
 use DateTime;
-use EasyRdf\Graph;
 use EasyRdf\Literal;
 use EasyRdf\Resource;
 
 final class WorkflowStatusEditor
 {
-    private Graph $graph;
-
     private array $workflowStatusMapping;
 
     private const PROPERTY_WORKFLOW_STATUS = 'udb:workflowStatus';
@@ -24,10 +21,8 @@ final class WorkflowStatusEditor
     private const PROPERTY_WORKFLOW_STATUS_DELETED = 'https://data.publiq.be/concepts/workflowStatus/deleted';
     private const PROPERTY_AVAILABLE_FROM = 'udb:availableFrom';
 
-    private function __construct(Graph $graph)
+    public function __construct()
     {
-        $this->graph = $graph;
-
         $this->workflowStatusMapping = [
             WorkflowStatus::DRAFT()->toString() => self::PROPERTY_WORKFLOW_STATUS_DRAFT,
             WorkflowStatus::READY_FOR_VALIDATION()->toString() => self::PROPERTY_WORKFLOW_STATUS_READY_FOR_VALIDATION,
@@ -35,68 +30,6 @@ final class WorkflowStatusEditor
             WorkflowStatus::REJECTED()->toString() => self::PROPERTY_WORKFLOW_STATUS_REJECTED,
             WorkflowStatus::DELETED()->toString() => self::PROPERTY_WORKFLOW_STATUS_DELETED,
         ];
-    }
-
-    public static function for(Graph $graph): self
-    {
-        return new self($graph);
-    }
-
-    public function draft(string $resourceIri): void
-    {
-        $resource = $this->graph->resource($resourceIri);
-
-        if (!$resource->hasProperty(self::PROPERTY_WORKFLOW_STATUS)) {
-            $resource->set(
-                self::PROPERTY_WORKFLOW_STATUS,
-                new Resource(self::PROPERTY_WORKFLOW_STATUS_DRAFT)
-            );
-        }
-    }
-
-    public function publish(string $resourceIri, string $publicationDate): void
-    {
-        $resource = $this->graph->resource($resourceIri);
-
-        $resource->set(
-            self::PROPERTY_WORKFLOW_STATUS,
-            new Resource(self::PROPERTY_WORKFLOW_STATUS_READY_FOR_VALIDATION)
-        );
-
-        $resource->set(
-            self::PROPERTY_AVAILABLE_FROM,
-            new Literal($publicationDate, null, 'xsd:dateTime')
-        );
-    }
-
-    public function approve(string $resourceIri): void
-    {
-        $resource = $this->graph->resource($resourceIri);
-
-        $resource->set(
-            self::PROPERTY_WORKFLOW_STATUS,
-            new Resource(self::PROPERTY_WORKFLOW_STATUS_APPROVED)
-        );
-    }
-
-    public function reject(string $resourceIri): void
-    {
-        $resource = $this->graph->resource($resourceIri);
-
-        $resource->set(
-            self::PROPERTY_WORKFLOW_STATUS,
-            new Resource(self::PROPERTY_WORKFLOW_STATUS_REJECTED)
-        );
-    }
-
-    public function delete(string $resourceIri): void
-    {
-        $resource = $this->graph->resource($resourceIri);
-
-        $resource->set(
-            self::PROPERTY_WORKFLOW_STATUS,
-            new Resource(self::PROPERTY_WORKFLOW_STATUS_DELETED)
-        );
     }
 
     public function setWorkflowStatus(Resource $resource, WorkflowStatus $workflowStatus): void
